requirements.txt
fastapi
uvicorn
jinja2
google-auth
pymongo
azure-storage-blob
python-multipart
